rspec: Remove Puppet::Util::Storage mock
In attempting to upgrade bodgit/puppet-postfix's[^1] dependency on this
module, I received the following error:

  Failure/Error: allow(Puppet::Util::Storage).to receive(:store)
    The use of doubles or partial doubles from rspec-mocks outside
    of the per-test lifecycle is not supported.

This appears to be the result of the change in
637108ebcdab86b2de29ff51b41dfc6db39ab4d9 to use a mock rather than a
stub. However, it is not clear to me why Puppet::Util::Storage needs to
be mocked, as all tests still pass in this module and puppet-postfix
without it, so remove.

[1]: https://github.com/bodgit/puppet-postfix